requirements.txt
Flask
praw
xai-sdk
opencv-python
insightface
numpy
Pillow
requests
onnxruntime
python-dotenv
gunicorn
